feat(lib): Moves all dto and rdo to lib
