+++ taskforce/apps/bff/src/main.ts
@@ -25,9 +25,7 @@ async function bootstrap() {
   SwaggerModule.setup('spec', app, document);
 
   await app.listen(port);
-  Logger.log(
-    `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`
-  );
+  Logger.log(`🚀 BFF is running on: http://localhost:${port}/${globalPrefix}`);
 }
 
 bootstrap();
