+++ taskforce/apps/tasks/src/app/comment/comment.const.ts
@@ -1,5 +1,2 @@
 export const COMMENT_NOT_FOUND_ERROR = 'comment not found';
-
-export const COMMENT_MIN_LENGTH = 10;
-export const COMMENT_MAX_LENGTH = 300;
 export const DEFAULT_COMMENTS_LIMIT = 50;
