+++ taskforce/apps/tasks/src/app/response/response.controller.ts
@@ -18,7 +18,7 @@ import {
   RolesGuard,
 } from '@taskforce/core';
 import { ResponseService } from './response.service';
-import { ResponseRdo } from './rdo/response.rdo';
+import { ResponseRdo } from '@taskforce/core';
 import { UserRoles } from '@taskforce/shared-types';
 
 @ApiTags('tasks')
