+++ taskforce/apps/tasks/src/app/review/review.const.ts
@@ -1,8 +1,3 @@
-export const TEXT_MAX_LENGTH = 500;
-export const TEXT_MIN_LENGTH = 50;
-export const RATING_MIN = 1;
-export const RATING_MAX = 5;
-
 export const REVIEW_ALREADY_EXISTS_ERROR =
   "Customer's review on this task already exists";
 export const WRONG_CUSTOMER_ERROR = "Can't add reviews to other users' tasks";
