+++ taskforce/apps/users/src/app/auth/auth.controller.ts
@@ -29,14 +29,16 @@ import {
   ApiConsumes,
 } from '@nestjs/swagger';
 import { AuthService } from './auth.service';
-import { ChangePasswordDto } from './dto/change-password.dto';
-import { CreateUserDto } from './dto/create-user.dto';
-import { LoginUserDto } from './dto/login-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { ContractorRdo } from './rdo/contractor.rdo';
-import { CustomerRdo } from './rdo/customer.rdo';
-import { LoggedInUserRdo } from './rdo/logged-in-user.rdo';
-import { UserRdo } from './rdo/user.rdo';
+import {
+  ChangePasswordDto,
+  CreateUserDto,
+  LoginUserDto,
+  UpdateUserDto,
+  ContractorRdo,
+  CustomerRdo,
+  LoggedInUserRdo,
+  UserRdo,
+} from '@taskforce/core';
 import { MongoIdValidationPipe } from '../pipes/mongo-id-validation.pipe';
 import { RtAuthGuard } from './guards';
 import { UserRoles } from '@taskforce/shared-types';
