+++ taskforce/libs/core/src/index.ts
@@ -14,3 +14,24 @@ export * from './lib/decorators/is-older-than.decorator';
 export * from './lib/decorators/is-date-greater-than.decorator';
 export * from './lib/utils/transform-tags';
 export * from './lib/utils/calculate-age';
+export * from './lib/dto/change-password.dto';
+export * from './lib/dto/change-task-status.dto';
+export * from './lib/dto/create-category.dto';
+export * from './lib/dto/create-comment.dto';
+export * from './lib/dto/create-review.dto';
+export * from './lib/dto/create-task.dto';
+export * from './lib/dto/create-user.dto';
+export * from './lib/dto/login-user.dto';
+export * from './lib/dto/update-task.dto';
+export * from './lib/dto/update-user.dto';
+export * from './lib/rdo/category.rdo';
+export * from './lib/rdo/comment.rdo';
+export * from './lib/rdo/contractor.rdo';
+export * from './lib/rdo/customer.rdo';
+export * from './lib/rdo/logged-in-user.rdo';
+export * from './lib/rdo/response.rdo';
+export * from './lib/rdo/review.rdo';
+export * from './lib/rdo/task-contractor.rdo';
+export * from './lib/rdo/task-customer.rdo';
+export * from './lib/rdo/task.rdo';
+export * from './lib/rdo/user.rdo';
