+++ taskforce/libs/core/src/lib/const/comment.const.ts
@@ -1,0 +1,2 @@
+export const COMMENT_MIN_LENGTH = 10;
+export const COMMENT_MAX_LENGTH = 300;
