+++ taskforce/libs/core/src/lib/const/review.const.ts
@@ -1,0 +1,4 @@
+export const TEXT_MAX_LENGTH = 500;
+export const TEXT_MIN_LENGTH = 50;
+export const RATING_MIN = 1;
+export const RATING_MAX = 5;
