+++ taskforce/libs/core/src/lib/const/task.const.ts
@@ -1,0 +1,15 @@
+export const DUE_DATE_LESS_THAN_CURRENT_ERROR =
+  'Task due date should be more than current date';
+export const TAG_FORMAT_NOT_VALID_ERROR =
+  'Tag should start with letter and not containe spaces';
+export const TASK_TITLE_MIN_LENGTH = 20;
+export const TASK_TITLE_MAX_LENGTH = 50;
+export const TASK_DESCRIPTION_MIN_LENGTH = 100;
+export const TASK_DESCRIPTION_MAX_LENGTH = 1024;
+export const TASK_ADDRESS_MIN_LENGTH = 10;
+export const TASK_ADDRESS_MAX_LENGTH = 255;
+export const TASK_PRICE_MIN_VALUE = 0;
+export const TASK_TAGS_MAX_NUM = 5;
+export const TAG_MIN_LENGTH = 3;
+export const TAG_MAX_LENGTH = 10;
+export const TAG_FORMAT_PATTERN = /^[A-Za-zА-Яа-я]\S+$/;
