+++ taskforce/libs/core/src/lib/dto/change-password.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString } from 'class-validator';
-import { PASSWORD_MAX_LENGTH, PASSWORD_MIN_LENGTH } from '../auth.const';
+import { PASSWORD_MAX_LENGTH, PASSWORD_MIN_LENGTH } from '../const/auth.const';
 
 export class ChangePasswordDto {
   @ApiProperty({
